// src/app/bill/models/bill-address-institution.ts
export class BillAddressInstitution {
    id: number;
    name: string;
    institution_id: number;
    institution: string;
    address1: string;
    address2: string;
    city: string;
    pincode: string;
    state_id: number;
    state: string;
    country: string;
    selected: boolean;
    country_id: number;
}